+++ chia/_tests/plotting/test_plot_manager.py
@@ -26,11 +26,9 @@ from chia.plotting.util import (
     PlotInfo,
     PlotRefreshEvents,
     PlotRefreshResult,
-    add_plot_directory,
-    get_plot_directories,
     remove_plot,
-    remove_plot_directory,
 )
+from chia.util.harvester_config import add_plot_directory, get_plot_directories, remove_plot_directory
 from chia.simulator.block_tools import get_plot_dir
 from chia.util.config import create_default_chia_config, lock_and_load_config, save_config
 from chia.util.streamable import VersionedBlob
