+++ chia/daemon/server.py
@@ -30,7 +30,7 @@ from chia.cmds.passphrase_funcs import default_passphrase, using_default_passphr
 from chia.daemon.keychain_server import KeychainServer, keychain_commands
 from chia.daemon.windows_signal import kill
 from chia.plotters.plotters import get_available_plotters
-from chia.plotting.util import add_plot_directory
+from chia.util.harvester_config import add_plot_directory
 from chia.server.server import ssl_context_for_server
 from chia.server.signal_handlers import SignalHandlers
 from chia.util.bech32m import encode_puzzle_hash
