+++ chia/plotting/util.py
@@ -15,6 +15,7 @@ from chia_rs.sized_ints import uint32
 from typing_extensions import final
 
 from chia.util.config import load_config, lock_and_load_config, save_config
+from chia.util.harvester_config import get_plot_directories
 from chia.util.streamable import Streamable, streamable
 
 log = logging.getLogger(__name__)
@@ -101,12 +102,6 @@ class HarvestingMode(IntEnum):
     GPU = 2
 
 
-def get_plot_directories(root_path: Path, config: dict | None = None) -> list[str]:
-    if config is None:
-        config = load_config(root_path, "config.yaml")
-    return config["harvester"]["plot_directories"] or []
-
-
 def get_plot_filenames(root_path: Path) -> dict[Path, list[Path]]:
     # Returns a map from directory to a list of all plots in the directory
     all_files: dict[Path, list[Path]] = {}
@@ -121,40 +116,6 @@ def get_plot_filenames(root_path: Path) -> dict[Path, list[Path]]:
             continue
         all_files[directory] = get_filenames(directory, recursive_scan, recursive_follow_links)
     return all_files
-
-
-def add_plot_directory(root_path: Path, str_path: str) -> dict:
-    path: Path = Path(str_path).resolve()
-    if not path.exists():
-        raise ValueError(f"Path doesn't exist: {path}")
-    if not path.is_dir():
-        raise ValueError(f"Path is not a directory: {path}")
-    log.debug(f"add_plot_directory {str_path}")
-    with lock_and_load_config(root_path, "config.yaml") as config:
-        if str(Path(str_path).resolve()) in get_plot_directories(root_path, config):
-            raise ValueError(f"Path already added: {path}")
-        if not config["harvester"]["plot_directories"]:
-            config["harvester"]["plot_directories"] = []
-        config["harvester"]["plot_directories"].append(str(Path(str_path).resolve()))
-        save_config(root_path, "config.yaml", config)
-    return config
-
-
-def remove_plot_directory(root_path: Path, str_path: str) -> None:
-    log.debug(f"remove_plot_directory {str_path}")
-    with lock_and_load_config(root_path, "config.yaml") as config:
-        str_paths: list[str] = get_plot_directories(root_path, config)
-        # If path str matches exactly, remove
-        if str_path in str_paths:
-            str_paths.remove(str_path)
-
-        # If path matches full path, remove
-        new_paths = [Path(sp).resolve() for sp in str_paths]
-        if Path(str_path).resolve() in new_paths:
-            new_paths.remove(Path(str_path).resolve())
-
-        config["harvester"]["plot_directories"] = [str(np) for np in new_paths]
-        save_config(root_path, "config.yaml", config)
 
 
 def remove_plot(path: Path):
